pyproject.toml
[project]
name = "building-effective-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.33.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "ruff>=0.15.12",
]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
